Update
Been working quite a bit part without commiting, bad reception on the train and worse memory when home.
Added a bundle of controllers for admin and the api.
Started working on the design. If we can call it design...
I have so far positioned the elements rather well and also manged to mess up the colour-theme properly.