Update
Finally made a design I can respect. It needs some smoothing animations and tweaks, but the general concept looks good.
The colour-scheme for admin and menus are still a work in progress and the footer contains close to zero information.
Need to implement posting of comments.

What has been done:
 - Design-idea: check
 - Vue for posts: check
 - API for posts: check
 - Vue for comments: partial check
 - API for comments: partial check